requirements.txt
asgiref==3.4.1
autopep8==1.5.7
Django==3.2.8
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
drf-nested-routers==0.93.4
flake8==4.0.1
flake8-html==0.4.1
importlib-metadata==4.8.1
Jinja2==3.0.2
MarkupSafe==2.0.1
mccabe==0.6.1
pycodestyle==2.8.0
pyflakes==2.4.0
Pygments==2.10.0
PyJWT==2.2.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
zipp==3.6.0
